pyproject.toml
[project]
name = "e-commerce-product-recommendation-system-using-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb==0.4.22",
    "fastapi==0.110.0",
    "google-generativeai>=0.8.6",
    "langchain==0.0.354",
    "langchain-community==0.0.20",
    "numpy==1.26.4",
    "pydantic==2.6.4",
    "pydantic-settings==2.2.1",
    "pypdf2==3.0.1",
    "python-docx==1.1.0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.9",
    "sentence-transformers==2.2.2",
    "uvicorn[standard]==0.27.1",
]
